`decoder`: Move `log::Log`-implementations into own modules
